+++ web/src/components/DepositCard/DepositCard.tsx
@@ -4,11 +4,7 @@ import { useContext, useEffect, useMemo, useState, useRef } from "react";
 import { Dec, DecUtils } from "@keplr-wallet/unit";
 import AnimatedArrowSpacer from "components/AnimatedDownArrowSpacer/AnimatedDownArrowSpacer";
 import Dropdown, { type DropdownOption } from "components/Dropdown/Dropdown";
-import {
-  type EvmChainInfo,
-  type IbcChainInfo,
-  toChainInfo,
-} from "config/chainConfigs";
+import type { EvmChainInfo, IbcChainInfo } from "config/chainConfigs";
 import { useConfig } from "config/hooks/useConfig";
 import { NotificationType } from "features/Notifications/components/Notification/types";
 import { NotificationsContext } from "features/Notifications/contexts/NotificationsContext";
@@ -16,7 +12,7 @@ import EthWalletConnector from "features/EthWallet/components/EthWalletConnector
 import { useEthWallet } from "features/EthWallet/hooks/useEthWallet";
 import { useEvmChainSelection } from "features/EthWallet/hooks/useEvmChainSelection";
 import { useIbcChainSelection } from "features/IbcChainSelector/hooks/useIbcChainSelection";
-import { getKeplrFromWindow, sendIbcTransfer } from "services/ibc";
+import { sendIbcTransfer } from "services/ibc";
 
 export default function DepositCard(): React.ReactElement {
   const { addNotification } = useContext(NotificationsContext);
@@ -48,16 +44,15 @@ export default function DepositCard(): React.ReactElement {
     selectedIbcCurrency,
     ibcBalance,
     isLoadingIbcBalance,
+    connectKeplrWallet,
   } = useIbcChainSelection(ibcChains);
-  const defaultIbcChainOption = useMemo(() => {
-    return ibcChainsOptions[0] || null;
-  }, [ibcChainsOptions]);
   const defaultIbcCurrencyOption = useMemo(() => {
     return ibcCurrencyOptions[0] || null;
   }, [ibcCurrencyOptions]);
 
   // selectedIbcChainOption allows us to ensure the label is set properly
-  // in the dropdown when connecting via additional action
+  // in the dropdown when connecting via an "additional option"s action,
+  //  e.g. the "Connect Keplr Wallet" option in the dropdown
   const selectedIbcChainOption = useMemo(() => {
     if (!selectedIbcChain) {
       return null;
@@ -88,24 +83,6 @@ export default function DepositCard(): React.ReactElement {
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [isAnimating, setIsAnimating] = useState<boolean>(false);
 
-  // create refs to hold the latest state values
-  const latestState = useRef({
-    evmUserAccount,
-    selectedWallet,
-    recipientAddress,
-    selectedEvmChain,
-  });
-
-  // update the ref whenever the state changes
-  useEffect(() => {
-    latestState.current = {
-      evmUserAccount,
-      selectedWallet,
-      recipientAddress,
-      selectedEvmChain,
-    };
-  }, [evmUserAccount, selectedWallet, recipientAddress, selectedEvmChain]);
-
   // check if form is valid whenever values change
   useEffect(() => {
     if (recipientAddress || amount) {
@@ -114,22 +91,6 @@ export default function DepositCard(): React.ReactElement {
     }
     checkIsFormValid(recipientAddress, amount);
   }, [recipientAddress, amount]);
-
-  // connect to keplr wallet when chain
-  useEffect(() => {
-    // FIXME - refactor this component to rely fully on useIbcChainSelection
-    if (!selectedIbcChain) {
-      return;
-    }
-    connectKeplrWallet().then((_) => {});
-  }, [selectedIbcChain]);
-
-  useEffect(() => {
-    if (!selectedEvmChain) {
-      return;
-    }
-    connectEVMWallet().then((_) => {});
-  }, [selectedEvmChain]);
 
   const updateAmount = (event: React.ChangeEvent<HTMLInputElement>) => {
     setAmount(event.target.value);
@@ -144,67 +105,32 @@ export default function DepositCard(): React.ReactElement {
     setIsRecipientAddressValid(addressValid);
   };
 
-  const connectKeplrWallet = async () => {
-    if (!selectedIbcChain) {
-      // select default chain if none selected, then return. effect handles retriggering.
-      selectIbcChain(defaultIbcChainOption.value);
+  // NOTE - this was required to ensure the latest state was used in a callback
+  //  used in the modal that connects to the evm wallet.
+  // create refs to hold the latest state values
+  const latestState = useRef({
+    evmUserAccount,
+    selectedWallet,
+    recipientAddress,
+    selectedEvmChain,
+  });
+  // update the ref whenever the state changes
+  useEffect(() => {
+    latestState.current = {
+      evmUserAccount,
+      selectedWallet,
+      recipientAddress,
+      selectedEvmChain,
+    };
+  }, [evmUserAccount, selectedWallet, recipientAddress, selectedEvmChain]);
+
+  // ensure evm wallet connection when selected EVM chain changes
+  useEffect(() => {
+    if (!selectedEvmChain) {
       return;
     }
-
-    // show a notification if the Keplr wallet is not installed
-    let keplr;
-    try {
-      keplr = getKeplrFromWindow();
-    } catch (err) {
-      addNotification({
-        toastOpts: {
-          toastType: NotificationType.DANGER,
-          component: (
-            <p>
-              Keplr wallet extension must be installed! You can find it{" "}
-              <a
-                target="_blank"
-                href="https://www.keplr.app/download"
-                rel="noreferrer"
-              >
-                here
-              </a>
-              .
-            </p>
-          ),
-          onAcknowledge: () => {},
-        },
-      });
-      return;
-    }
-
-    // suggest the chain if it doesn't exist in Keplr
-    try {
-      await keplr.getKey(selectedIbcChain.chainId);
-    } catch (e) {
-      if (
-        e instanceof Error &&
-        (e.message.startsWith("There is no chain info") ||
-          e.message.startsWith("There is no modular chain info"))
-      ) {
-        try {
-          await keplr.experimentalSuggestChain(toChainInfo(selectedIbcChain));
-        } catch (e) {
-          if (e instanceof Error) {
-            selectIbcChain(null);
-          }
-        }
-      } else {
-        addNotification({
-          toastOpts: {
-            toastType: NotificationType.DANGER,
-            message: "Failed to get key from Keplr wallet.",
-            onAcknowledge: () => {},
-          },
-        });
-      }
-    }
-  };
+    connectEVMWallet().then((_) => {});
+  }, [selectedEvmChain]);
 
   const connectEVMWallet = async () => {
     if (!selectedEvmChain) {
@@ -279,7 +205,8 @@ export default function DepositCard(): React.ReactElement {
           addNotification({
             toastOpts: {
               toastType: NotificationType.DANGER,
-              message: "Failed to get account from Keplr wallet. Does this address have funds for the selected chain?",
+              message:
+                "Failed to get account from Keplr wallet. Does this address have funds for the selected chain?",
               onAcknowledge: () => {},
             },
           });
@@ -300,6 +227,7 @@ export default function DepositCard(): React.ReactElement {
     }
   };
 
+  // this additional option is the "Connect Keplr Wallet" button
   const additionalIbcOptions = useMemo(
     () => [
       {
@@ -313,6 +241,7 @@ export default function DepositCard(): React.ReactElement {
     [connectKeplrWallet],
   );
 
+  // this additional option is the "Connect EVM Wallet" button
   const additionalEvmOptions = useMemo(() => {
     return [
       {
